+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- discovery functions
+
 ## 0.4.14 2019-02-08
 
 ### Fixed